Revert "Refactor sale model to remove duplicate code"
