--- src/main/java/bg/softuni/invoice/model/base/SaleBaseModel.java
@@ -1,34 +1,0 @@
-package bg.softuni.invoice.model.base;
-
-import java.math.BigDecimal;
-
-public abstract class SaleBaseModel {
-
-    private String name;
-    private int quantity;
-    private BigDecimal price;
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public int getQuantity() {
-        return quantity;
-    }
-
-    public void setQuantity(int quantity) {
-        this.quantity = quantity;
-    }
-
-    public BigDecimal getPrice() {
-        return price;
-    }
-
-    public void setPrice(BigDecimal price) {
-        this.price = price;
-    }
-}
