+++ src/main/java/bg/softuni/invoice/model/service/SaleServiceModel.java
@@ -1,11 +1,39 @@
 package bg.softuni.invoice.model.service;
 
-import bg.softuni.invoice.model.base.SaleBaseModel;
 import bg.softuni.invoice.model.enumerated.VatValue;
 
-public class SaleServiceModel extends SaleBaseModel {
+import java.math.BigDecimal;
 
+public class SaleServiceModel {
+
+    private String name;
+    private int quantity;
+    private BigDecimal price;
     private VatValue vatValue;
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public int getQuantity() {
+        return quantity;
+    }
+
+    public void setQuantity(int quantity) {
+        this.quantity = quantity;
+    }
+
+    public BigDecimal getPrice() {
+        return price;
+    }
+
+    public void setPrice(BigDecimal price) {
+        this.price = price;
+    }
 
     public VatValue getVatValue() {
         return vatValue;
